Remove other classes
