+++ lib/dashboard/advice/advice_electric_annual.rb
@@ -1,21 +1,4 @@
-require_relative './advice_general.rb'
-class AdviceElectricityAnnual < AdviceBenchmark
-  def aggregate_meter
-    @school.aggregated_electricity_meters
-  end
-
-  def normalised_benchmark_chart_name
-    :benchmark_electric_only_£_varying_floor_area_pupils
-  end
-
-  private
-
-  def valid_meters
-    [@school.aggregated_electricity_meters]
-  end
-end
-
-# require_relative '../charting_and_reports/old_advice/dashboard_analysis_advice.rb'
+require_relative '../charting_and_reports/old_advice/dashboard_analysis_advice.rb'
 
 # legacy energy analysis advice: provides text either side
 # of the :group_by_week_electricity_versus_benchmark chart
