--- lib/dashboard/advice/advice_storage_heaters.rb
@@ -1,23 +1,0 @@
-class AdviceStorageHeaters < AdviceBenchmark
-  def relevance
-    @school.storage_heaters? ? :relevant : :never_relevant
-  end
-
-  def rating
-    5.0
-  end
-
-  def aggregate_meter
-    @school.storage_heater_meter
-  end
-
-  def normalised_benchmark_chart_name
-    :benchmark_storage_heater_only_£_varying_floor_area_pupils
-  end
-
-  private
-
-  def valid_meters
-    [@school.storage_heater_meter]
-  end    
-end
