+++ lib/dashboard/charting_and_reports/old_advice/dashboard_analysis_advice.rb
@@ -177,7 +177,6 @@ class DashboardChartAdviceBase
       res = DashboardEnergyAdvice.solar_pv_advice_factory(chart_type, school, chart_definition, chart_data, chart_symbol) if res.nil?
       res = DashboardEnergyAdvice.storage_heater_advice_factory(chart_type, school, chart_definition, chart_data, chart_symbol) if res.nil?
       res = DashboardEnergyAdvice.co2_advice_factory(chart_type, school, chart_definition, chart_data, chart_symbol) if res.nil?
-      res = DashboardEnergyAdvice.financial_advice_factory(chart_type, school, chart_definition, chart_data, chart_symbol) if res.nil?
       res
     end
   end
